LayoutTests/media/media-source/media-source-real-remove-too-much.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-remove-too-much</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Real-MP4 equivalent of media-source-remove-too-much.html
    // Tests that remove(0, 4) removes samples from 0 up to the sync frame at PTS=4.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        run('video.src = URL.createObjectURL(source)');
        await waitFor(source, 'sourceopen');

        run(`sourceBuffer = source.addSourceBuffer('${MP4.VIDEO_TYPE}')`);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // 9 samples, sync at 0, 4, 8
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
            ]}]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        run('sourceBuffer.remove(0, 4)');
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.start(0)', 4);

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <div>Tests SourceBuffer.remove() with real MP4 data. Removes [0, 4) and verifies
    removal extends to the sync frame boundary at PTS=4.</div>
    <video></video>
</body>
</html>
